Update SDAtom-WebUi-us.user.js
Added back in Regional Prompter.

Forgot to mention Detail Daemon as well

I will need to rework this repo to have it better set up.

Base version for older forge and reforge without the extensions.

and ones with the extensions. I could be annoying for some who don't need extensions included and then have the errors pop up in the log.